backend/requirements.txt
Flask==2.3.4
flask-cors==3.0.10
numpy==1.26.3
joblib==1.3.2
scikit-learn==1.3.2
tensorflow==2.12.0
Pillow==10.2.0
